backend/requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase==2.0.2
psycopg2-binary==2.9.9

# Background Processing (Simple async)
asyncio

# HTTP Client
httpx>=0.24.0,<0.25.0
aiofiles==23.2.1

# Image Processing
Pillow==10.1.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
